backend/requirements.txt
openai==1.14.2
flask==3.0.2
flask_cors==3.0.10
python-dotenv==1.0.1
google-api-python-client==2.122.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gunicorn==22.0.0
google-cloud-storage==2.16.0
Pillow==10.3.0